Switch to universal rendering pipeline
Materials had to be recalculated and some reapplied.
With probuilder new objects automatically should get the right standard
material. If not go to Edit->Preferences->probuilder
And set the default material to be
Samples->Probuilder-><Version>->UniversalRenderPipelineSupport->Material->Pro..*
On old gameobjects with the old probuilder material, you might
have to apply the material in the mesh renderer tab.
Select multiple objects, drag the material in and
everything should be applied with old settings like color remaining.

The pipeline is first of all good practice to use (e.g. for performance)
and secondly the minimum neccesary for advanced features like shader graph editing.
Enjoy!